Re-organize pipeline "processors" into common package
