+++ pkg/pipeline/processors/aws_s3_archive_uploader.go
@@ -1,4 +1,4 @@
-package output_processors
+package processors
 
 import (
     "os"
