+++ pkg/pipeline/processors/filesystem_cache_writer.go
@@ -1,4 +1,4 @@
-package output_processors
+package processors
 
 import (
     "archive/tar"
