+++ pkg/pipeline/processors/formatter.go
@@ -1,4 +1,4 @@
-package transformers
+package processors
 
 import (
     "fireside/pkg/pipeline/data"
